refactor: Move p2p protocols to network module
