+++ Cargo.toml
@@ -62,9 +62,6 @@ members = [
     "benches",
     "error",
     "notify",
-    "protocols/ping",
-    "protocols/identify",
-    "protocols/discovery",
 ]
 
 [profile.release]
