+++ network/Cargo.toml
@@ -11,14 +11,11 @@ serde = { version = "1.0", features = ["derive"] }
 ckb-util = { path = "../util" }
 ckb-stop-handler = { path = "../util/stop-handler" }
 ckb-logger = { path = "../util/logger" }
-tokio = { version = "0.2.11", features = ["time", "io-util", "tcp", "dns", "rt-threaded", "blocking"] }
+tokio = { version = "0.2.11", features = ["time", "io-util", "tcp", "dns", "rt-threaded", "blocking", "stream"] }
 tokio-util = { version = "0.2.0", features = ["codec"] }
 futures = "0.3"
 crossbeam-channel = "0.3"
 p2p = { version="0.3.0-alpha.2", package="tentacle", features = ["molc"] }
-p2p-ping = { package="ckb-ping", path = "../protocols/ping" }
-p2p-discovery = { package="ckb-discovery", path = "../protocols/discovery" }
-p2p-identify = { package="ckb-identify", path = "../protocols/identify" }
 faketime = "0.2.0"
 lazy_static = "1.3.0"
 bs58 = "0.3.0"
